Fix PWM speed control

+++ pico/motor/motor.h
@@ -12,9 +12,9 @@
 
 void MOTOR_init(void);
 
-void MOTOR_setSpeed(uint16_t dutyCycle, int motor);
+void MOTOR_setSpeed(uint8_t dutyCycle, int motor);
 void MOTOR_stop(int motor);
-uint16_t MOTOR_getSpeed(int motor);
+uint8_t MOTOR_getSpeed(int motor);
 
 void MOTOR_setDirection(int dir, int motor);
 int MOTOR_getDirection(int motor);
